requirements.txt
# Deep Learning Framework
torch==2.0.0+cu117
torchvision==0.15.0+cu117
timm==1.0.3
einops==0.8.0
safetensors==0.4.3

# Medical Image Processing
SimpleITK==2.3.1
nibabel==5.1.0
monai==1.3.1
opencv-python==4.9.0.80

# Machine Learning & Data Processing
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
umap-learn==0.5.6

# Visualization and Progress
matplotlib==3.7.5
tqdm==4.66.5

# AI Model Hub
huggingface-hub==0.26.2

# Data Augmentation and Utils
albumentations==1.4.7
PyYAML==6.0.1
icecream==2.1.3
